add package for userstats document. remove enclosing class for last usage
