+++ src/test/java/it/pagopa/ecommerce/commons/utils/UserStatisticsTestUtils.java
@@ -1,12 +1,13 @@
 package it.pagopa.ecommerce.commons.utils;
 
-import it.pagopa.ecommerce.commons.documents.UserStatistics;
+import it.pagopa.ecommerce.commons.documents.userstats.v1.LastUsage;
+import it.pagopa.ecommerce.commons.documents.userstats.v1.UserStatistics;
 
 import java.time.Instant;
 import java.util.UUID;
 
-import static it.pagopa.ecommerce.commons.documents.UserStatistics.LastUsage.PaymentType.GUEST_PAYMENT_METHOD;
-import static it.pagopa.ecommerce.commons.documents.UserStatistics.LastUsage.PaymentType.SAVED_WALLET;
+import static it.pagopa.ecommerce.commons.documents.userstats.v1.LastUsage.PaymentType.GUEST_PAYMENT_METHOD;
+import static it.pagopa.ecommerce.commons.documents.userstats.v1.LastUsage.PaymentType.SAVED_WALLET;
 
 public class UserStatisticsTestUtils {
 
@@ -14,15 +15,15 @@ public class UserStatisticsTestUtils {
         return new UserStatistics("userId", lastUsageWallet());
     }
 
-    private static UserStatistics.LastUsage lastUsageWallet() {
-        return new UserStatistics.LastUsage(SAVED_WALLET, UUID.randomUUID(), Instant.now());
+    private static LastUsage lastUsageWallet() {
+        return new LastUsage(SAVED_WALLET, UUID.randomUUID(), Instant.now());
     }
 
     public static UserStatistics userStatisticsGuest() {
         return new UserStatistics("userId", lastUsageGuest());
     }
 
-    private static UserStatistics.LastUsage lastUsageGuest() {
-        return new UserStatistics.LastUsage(GUEST_PAYMENT_METHOD, UUID.randomUUID(), Instant.now());
+    private static LastUsage lastUsageGuest() {
+        return new LastUsage(GUEST_PAYMENT_METHOD, UUID.randomUUID(), Instant.now());
     }
 }
